Fix Primitives' transformation
The primitives had different behavior under the transformations. Rotation and Scale for the triangle are applied relative to the origin of the WCS, whereas for the plane, sphere, and disc - relative to the origin of the primitives. This commit changes that for sphere, plane, and triangle - now all the transformations are applied relative to the origin of the WCS.